+++ server/src/services/mod.rs
@@ -1,7 +1,8 @@
 //! # Services
 //!
 mod auth;
+mod clients;
 mod csrf;
 mod metrics;
 
-pub use crate::services::{auth::*, csrf::*, metrics::*};
+pub use crate::services::{auth::*, clients::*, csrf::*, metrics::*};