+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Move example endpoints into their own gRPC service
 -   Move authentication example code into server Auth service module
+-   Add server Clients service module and example endpoint
 
 ## [0.3.3] - 2021-04-18